Edit to fix container div issue that was causing conflict with Twitter Bootstrap
