+++ assets/components/syntaxhighlighter/scripts/shAutoloader.js
@@ -1,17 +1,114 @@
+(function() {
+
+var sh = SyntaxHighlighter;
+
 /**
- * SyntaxHighlighter
- * http://alexgorbatchev.com/SyntaxHighlighter
+ * Provides functionality to dynamically load only the brushes that a needed to render the current page.
  *
- * SyntaxHighlighter is donationware. If you are using it, please donate.
- * http://alexgorbatchev.com/SyntaxHighlighter/donate.html
- *
- * @version
- * 3.0.83 (July 02 2010)
+ * There are two syntaxes that autoload understands. For example:
  * 
- * @copyright
- * Copyright (C) 2004-2010 Alex Gorbatchev.
+ * SyntaxHighlighter.autoloader(
+ *     [ 'applescript',          'Scripts/shBrushAppleScript.js' ],
+ *     [ 'actionscript3', 'as3', 'Scripts/shBrushAS3.js' ]
+ * );
  *
- * @license
- * Dual licensed under the MIT and GPL licenses.
+ * or a more easily comprehendable one:
+ *
+ * SyntaxHighlighter.autoloader(
+ *     'applescript       Scripts/shBrushAppleScript.js',
+ *     'actionscript3 as3 Scripts/shBrushAS3.js'
+ * );
  */
-eval(function(p,a,c,k,e,d){e=function(c){return(c<a?'':e(parseInt(c/a)))+((c=c%a)>35?String.fromCharCode(c+29):c.toString(36))};if(!''.replace(/^/,String)){while(c--){d[e(c)]=k[c]||e(c)}k=[function(e){return d[e]}];e=function(){return'\\w+'};c=1};while(c--){if(k[c]){p=p.replace(new RegExp('\\b'+e(c)+'\\b','g'),k[c])}}return p}('(2(){1 h=5;h.I=2(){2 n(c,a){4(1 d=0;d<c.9;d++)i[c[d]]=a}2 o(c){1 a=r.H("J"),d=3;a.K=c;a.M="L/t";a.G="t";a.u=a.v=2(){6(!d&&(!8.7||8.7=="F"||8.7=="z")){d=q;e[c]=q;a:{4(1 p y e)6(e[p]==3)B a;j&&5.C(k)}a.u=a.v=x;a.D.O(a)}};r.N.R(a)}1 f=Q,l=h.P(),i={},e={},j=3,k=x,b;5.T=2(c){k=c;j=q};4(b=0;b<f.9;b++){1 m=f[b].w?f[b]:f[b].S(/\\s+/),g=m.w();n(m,g)}4(b=0;b<l.9;b++)6(g=i[l[b].E.A]){e[g]=3;o(g)}}})();',56,56,'|var|function|false|for|SyntaxHighlighter|if|readyState|this|length|||||||||||||||||true|document||javascript|onload|onreadystatechange|pop|null|in|complete|brush|break|highlight|parentNode|params|loaded|language|createElement|autoloader|script|src|text|type|body|removeChild|findElements|arguments|appendChild|split|all'.split('|'),0,{}))
+sh.autoloader = function()
+{
+	var list = arguments,
+		elements = sh.findElements(),
+		brushes = {},
+		scripts = {},
+		all = SyntaxHighlighter.all,
+		allCalled = false,
+		allParams = null,
+		i
+		;
+		
+	SyntaxHighlighter.all = function(params)
+	{
+		allParams = params;
+		allCalled = true;
+	};
+	
+	function addBrush(aliases, url)
+	{
+		for (var i = 0; i < aliases.length; i++)
+			brushes[aliases[i]] = url;
+	};
+	
+	function getAliases(item)
+	{
+		return item.pop
+			? item
+			: item.split(/\s+/)
+			;
+	}
+	
+	// create table of aliases and script urls
+	for (i = 0; i < list.length; i++)
+	{
+		var aliases = getAliases(list[i]),
+			url = aliases.pop()
+			;
+			
+		addBrush(aliases, url);
+	}
+	
+	// dynamically add <script /> tags to the document body
+	for (i = 0; i < elements.length; i++)
+	{
+		var url = brushes[elements[i].params.brush];
+		
+		if (!url)
+			continue;
+		
+		scripts[url] = false;
+		loadScript(url);
+	}
+	
+	function loadScript(url)
+	{
+		var script = document.createElement('script'),
+			done = false
+			;
+		
+		script.src = url;
+		script.type = 'text/javascript';
+		script.language = 'javascript';
+		script.onload = script.onreadystatechange = function()
+		{
+			if (!done && (!this.readyState || this.readyState == 'loaded' || this.readyState == 'complete'))
+			{
+				done = true;
+				scripts[url] = true;
+				checkAll();
+				
+				// Handle memory leak in IE
+				script.onload = script.onreadystatechange = null;
+				script.parentNode.removeChild(script);
+			}
+		};
+		
+		// sync way of adding script tags to the page
+		document.body.appendChild(script);
+	};
+	
+	function checkAll()
+	{
+		for(var url in scripts)
+			if (scripts[url] == false)
+				return;
+		
+		if (allCalled)
+			SyntaxHighlighter.highlight(allParams);
+	};
+};
+
+})();
