+++ assets/components/syntaxhighlighter/scripts/shLegacy.js
@@ -1,17 +1,141 @@
-/**
- * SyntaxHighlighter
- * http://alexgorbatchev.com/SyntaxHighlighter
- *
- * SyntaxHighlighter is donationware. If you are using it, please donate.
- * http://alexgorbatchev.com/SyntaxHighlighter/donate.html
- *
- * @version
- * 3.0.83 (July 02 2010)
- * 
- * @copyright
- * Copyright (C) 2004-2010 Alex Gorbatchev.
- *
- * @license
- * Dual licensed under the MIT and GPL licenses.
- */
-eval(function(p,a,c,k,e,d){e=function(c){return(c<a?'':e(parseInt(c/a)))+((c=c%a)>35?String.fromCharCode(c+29):c.toString(36))};if(!''.replace(/^/,String)){while(c--){d[e(c)]=k[c]||e(c)}k=[function(e){return d[e]}];e=function(){return'\\w+'};c=1};while(c--){if(k[c]){p=p.replace(new RegExp('\\b'+e(c)+'\\b','g'),k[c])}}return p}('3 u={8:{}};u.8={A:4(c,k,l,m,n,o){4 d(a,b){2 a!=1?a:b}4 f(a){2 a!=1?a.E():1}c=c.I(":");3 g=c[0],e={};t={"r":K};M=1;5=8.5;9(3 j R c)e[c[j]]="r";k=f(d(k,5.C));l=f(d(l,5.D));m=f(d(m,5.s));o=f(d(o,5.Q));n=f(d(n,5["x-y"]));2{P:g,C:d(t[e.O],k),D:d(t[e.N],l),s:d({"r":r}[e.s],m),"x-y":d(4(a,b){9(3 h=T S("^"+b+"\\\\[(?<q>\\\\w+)\\\\]$","U"),i=1,p=0;p<a.7;p++)6((i=h.J(a[p]))!=1)2 i.q;2 1}(c,"G"),n)}},F:4(c,k,l,m,n,o){4 d(){9(3 a=H,b=0;b<a.7;b++)6(a[b]!==1){6(z a[b]=="L"&&a[b]!="")2 a[b]+"";6(z a[b]=="X"&&a[b].q!="")2 a[b].q+""}2 1}4 f(a,b,h){h=12.13(h);9(3 i=0;i<h.7;i++)h[i].V("15")==b&&a.Y(h[i])}3 g=[];f(g,c,"Z");f(g,c,"W");6(g.7!==0)9(c=0;c<g.7;c++){3 e=g[c],j=d(e.B["14"],e.10,e.B.v,e.v);6(j!==1){j=u.8.A(j,k,l,m,n,o);8.11(j,e)}}}};',62,68,'|null|return|var|function|defaults|if|length|SyntaxHighlighter|for|||||||||||||||||value|true|collapse|reverse|dp|language||first|line|typeof|parseParams|attributes|gutter|toolbar|toString|HighlightAll|firstline|arguments|split|exec|false|string|result|nocontrols|nogutter|brush|ruler|in|XRegExp|new|gi|getAttribute|textarea|object|push|pre|className|highlight|document|getElementsByTagName|class|name'.split('|'),0,{}))
+var dp = {
+	SyntaxHighlighter : {}
+};
+
+dp.SyntaxHighlighter = {
+	parseParams: function(
+						input,
+						showGutter, 
+						showControls, 
+						collapseAll, 
+						firstLine, 
+						showColumns
+						)
+	{
+		function getValue(list, name)
+		{
+			var regex = new XRegExp('^' + name + '\\[(?<value>\\w+)\\]$', 'gi'),
+				match = null
+				;
+			
+			for (var i = 0; i < list.length; i++) 
+				if ((match = regex.exec(list[i])) != null)
+					return match.value;
+			
+			return null;
+		};
+		
+		function defaultValue(value, def)
+		{
+			return value != null ? value : def;
+		};
+		
+		function asString(value)
+		{
+			return value != null ? value.toString() : null;
+		};
+
+		var parts = input.split(':'),
+			brushName = parts[0],
+			options = {},
+			straight = { 'true' : true }
+			reverse = { 'true' : false },
+			result = null,
+			defaults = SyntaxHighlighter.defaults
+			;
+		
+		for (var i in parts)
+			options[parts[i]] = 'true';
+
+		showGutter = asString(defaultValue(showGutter, defaults.gutter));
+		showControls = asString(defaultValue(showControls, defaults.toolbar));
+		collapseAll = asString(defaultValue(collapseAll, defaults.collapse)); 
+		showColumns = asString(defaultValue(showColumns, defaults.ruler));
+		firstLine = asString(defaultValue(firstLine, defaults['first-line'])); 
+
+		return {
+			brush			: brushName,
+			gutter			: defaultValue(reverse[options.nogutter], showGutter),
+			toolbar			: defaultValue(reverse[options.nocontrols], showControls),
+			collapse		: defaultValue(straight[options.collapse], collapseAll),
+			// ruler			: defaultValue(straight[options.showcolumns], showColumns),
+			'first-line'	: defaultValue(getValue(parts, 'firstline'), firstLine)
+		};
+	},
+	
+	HighlightAll: function(
+						name, 
+						showGutter /* optional */, 
+						showControls /* optional */, 
+						collapseAll /* optional */, 
+						firstLine /* optional */, 
+						showColumns /* optional */
+						)
+	{
+		function findValue()
+		{
+			var a = arguments;
+			
+			for (var i = 0; i < a.length; i++) 
+			{
+				if (a[i] === null) 
+					continue;
+				
+				if (typeof(a[i]) == 'string' && a[i] != '') 
+					return a[i] + '';
+				
+				if (typeof(a[i]) == 'object' && a[i].value != '') 
+					return a[i].value + '';
+			}
+			
+			return null;
+		};
+
+		function findTagsByName(list, name, tagName)
+		{
+			var tags = document.getElementsByTagName(tagName);
+			
+			for (var i = 0; i < tags.length; i++) 
+				if (tags[i].getAttribute('name') == name) 
+					list.push(tags[i]);
+		}
+		
+		var elements = [],
+			highlighter = null,
+			registered = {},
+			propertyName = 'innerHTML'
+			;
+		
+		// for some reason IE doesn't find <pre/> by name, however it does see them just fine by tag name... 
+		findTagsByName(elements, name, 'pre');
+		findTagsByName(elements, name, 'textarea');
+
+		if (elements.length === 0)
+			return;
+		
+		for (var i = 0; i < elements.length; i++)
+		{
+			var element = elements[i],
+				params = findValue(
+					element.attributes['class'], element.className, 
+					element.attributes['language'], element.language
+					),
+				language = ''
+				;
+			
+			if (params === null) 
+				continue;
+
+			params = dp.SyntaxHighlighter.parseParams(
+				params,
+				showGutter, 
+				showControls, 
+				collapseAll, 
+				firstLine, 
+				showColumns
+				);
+
+			SyntaxHighlighter.highlight(params, element);
+		}
+	}
+};
