+++ core/components/syntaxhighlighter/elements/snippets/syntaxhighlighter.snippet.php
@@ -43,6 +43,7 @@ $brushArray = explode(',', $brushes);
 
 $baseURL = $modx->getOption('sh.assets_url', null, $modx->getOption('assets_url') . 'components/syntaxhighlighter/');
 
+$modx->regClientStartupScript($baseURL . 'scripts/XRegExp.js');
 $modx->regClientStartupScript($baseURL . 'scripts/shCore.js');
 
 foreach($brushArray as $brush) {
